Remove Ajax implementation in Android module Issue: 88024
